fh-gateway/requirements.txt
appdirs==1.4.4
attrs==20.3.0
Authlib==0.15.3
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.4.4
dacite==1.6.0
ecdsa==0.14.1
fastapi==0.63.0
fastapi-cloudauth==0.2.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
lovely.testlayers==0.7.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
regex==2020.11.13
requests==2.25.1
responses==0.12.1
rfc3986==1.4.0
rsa==4.7
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
-e ./fh-gateway
